AI/requirements.txt
# AI 팟캐스트 생성 시스템 의존성

# 기본 라이브러리
requests>=2.32.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# LangChain 관련
langchain-core>=0.1.0
langchain-google-genai>=1.0.0

# Google Cloud TTS
google-cloud-texttospeech>=2.16.0

# 오디오 처리
pydub>=0.25.1

# 데이터 처리
pandas>=2.0.0
numpy>=1.24.0

# 유틸리티
pathlib>=1.0.1
typing-extensions>=4.0.0

# 개발 및 테스트 (선택사항)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
